requirements.txt
torch>=2.2.0
torchvision>=0.17.0
ultralytics>=8.3.30
opencv-python>=4.10.0.84
numpy>=1.26.0
pandas>=2.2.0
scikit-learn>=1.5.0
scipy>=1.11.0
Pillow>=10.3.0
PyYAML>=6.0.1
tqdm>=4.66.4
matplotlib>=3.8.0
seaborn>=0.13.0
albumentations>=1.4.0
einops>=0.8.0
torchcam>=0.4.0
folium>=0.16.0
streamlit>=1.37.0
streamlit-folium>=0.20.0
piexif>=1.1.3
exifread>=3.0.0



